+++ dashboard/src/app/(authenticated)/[tenantId]/(diagram)/components/Diagram.tsx
@@ -13,7 +13,7 @@ import ReactFlow, { Controls, useEdgesState, useNodesState } from 'reactflow'
 import 'reactflow/dist/base.css'
 import { ThreadProvider, ThreadType } from '../context'
 import { edgeTypes } from './EdgeTypes'
-import { extractAllEdges } from './EdgeTypes/extractAllEdges'
+import { extractAllEdges } from './EdgeTypes/extractEdges'
 import { Layouter } from './Layouter'
 import nodeTypes from './NodeTypes'
 import { extractAllNodes } from './NodeTypes/extractNodes'
