+++ dashboard/src/app/(authenticated)/[tenantId]/(diagram)/components/EdgeTypes/extractEdges.ts
